pyproject.toml
[tool.poetry]
name = "cncrd-ext-player"
version = "3.0.0"
description = "Player extension (music-related functionality) for Concord"
authors = ["Nariman Safiulin <woofilee@gmail.com>"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/Roolat/concord-ext-player"

keywords = ["discord", "concord", "bot", "player", "music"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: Russian",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include = "concord" },
]

[tool.poetry.dependencies]
python = "^3.6"
cncrd = "^1.0.0"
"cncrd-ext-audio" = { git = "https://github.com/Roolat/concord-ext-audio.git", branch = "dev" }
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", branch = "rewrite", extras = ["voice"] }
streamlink = "^0.14.0"
youtube-dl = "*"

[tool.poetry.dev-dependencies]
black = "^18.9b0"
pre-commit = "^1.10"

[tool.black]
line-length = 80
py36 = true
